django_app/requirements.txt
Django>=5.0
djangorestframework>=3.15
httpx>=0.27
requests>=2.32
python-dotenv>=1.0
uvicorn>=0.30

# LangGraph + LangChain + Ollama + MCP adapter
langgraph>=0.2.50
langchain>=0.3.0
langchain-core>=0.3.0
langchain-ollama>=0.2.0
langchain-mcp-adapters>=0.1.0
